OPTM-75 Better support for multiple locations for one peer
- Move location for last successful peer connection to the front
- Some refactoring
- Moved tests/network/web_socket_peer_test.dart to tests/services/peer_service_test.dart because it mostly tests the PeerService
- Added some tests for peers with multiple peer locations
